testdata/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
lz4==3.1.0                # via -r requirements.in
mimesis==4.0.0            # via -r requirements.in
mysql-connector-python==8.0.21  # via -r requirements.in
protobuf==3.12.4          # via mysql-connector-python
six==1.15.0               # via protobuf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
